Add  function to render module and update test cases accordingly
